Add Uber logo to Splash screen using FastImage
This commit adds the Uber logo to the Splash screen using the FastImage component for better performance. The white version of the logo is used and added to the Images utility file. The background color of the Splash screen is changed to black. Additionally, the RNFastImage package is added to the project dependencies.