feat: Add SafeAreaLayout component and implement it in HomeScreen
This commit adds a new component called SafeAreaLayout which wraps the children components and sets the background color and status bar style. It also implements the SafeAreaLayout component in the HomeScreen component. Additionally, the AuthStack and HomeStack navigators have been updated to set the initial route and hide the header. Some new types have also been added to the types file.